Examples/essence-selfhosted/requirements.txt
python-dotenv~=1.1
bithuman>=1.7.0
livekit-agents[openai,bithuman,silero]~=1.4
loguru>=0.7.0
sounddevice>=0.4
opencv-python>=4.8
numpy>=1.24
